Change order of object combination
This seems to make more sense when changing things like number of hits
returned on subsequent "search" calls after the first init.

Let me know if this breaks an intended logic, but it seems to me that if
I explicitly supply settings to a search then they should be used over
the defaults or init variables.