remove json2.js, use an npm module, IE7 compat
We cannot run the testing suite on IE7 because it's a really old browser
But we can run the examples and they are working for the simple ones.

It means algoliasearch works on IE7 for the big parts.

We migth get rid or stop supporting it if it requires too much work