Fix #251: increments hostIndex (host adress) on fallback
This issue was caused because we do not distinguish CORS errors from DNS
resolution one and the code for incrementing the host index was moved into
the part that switches to the fallback.

Truth is that the current solution is not optimal as we are switching to
JSONP when we can't resolve the server name.