feat(copyIndex): back-compatible add of scope (#618)
copyIndex can now accept a third parameter (scope) which is an array with values 'settings', 'synonyms', 'rules'. An empty array is treated as not given, since the api otherwise returns an error, and I have the feeling that this might be what people expect.

You can still have the callback as third argument if you want, or leave the third undefined and put it in the fourth position.

I added new tests with a valid array, empty array and undefined.

Wanted to test if callback is accepted on both position 3 and 4, but that seems to be quite hard, so for now I'll just let this be.