feat(security): use https by default
Originally the fact to use the page protocol or http by default was made
to speed up first requests to avoid initiating an https connection.

Nowadays this is no more of a big issue performance wise (citation
needed), but security wise, the use of http by default combined with
JSONP can be a big issue of MITM attacks.