feat(cache): cache the requests instead of responses (#694)
If you pass `_useRequestCache: true` as an option to the client then the cache behaviour will change. If you launch two requests at the same time then only one request will be done while the two callbacks/promises will still be called/resolved.

Previously, the client would have made two requests instead of one.

This is implemented by filling the cache as soon as the request launches instead of waiting for the response to fill it.