test(bundlesize): run build before testing size
The reason is pretty simple: since we track the `dist` in GitHub, if we don't do a build before testing the build size; we are actually testing the size of the previous build, rather than the current build.

Also I realised that the `yarn test` command was not updated with previous test name changes, so did that.

The bundlesize had to be increased for the lite build, since #694 increased that size over the limit.