fix(search): add opts.strategy to request object
Up until now this was added to the URL query parameters, but according to the [docs](https://www.algolia.com/doc/rest-api/search#method-param-strategy), this is a parameter in the body. Because of this, the strategy cannot work in JSONP anymore, but it also seems to never have worked in the first place, so I'm not too concerned about that.

fixes #719