feat(cache): allow export & import
In SSR, you usually want to do the same query on the server as on the client, however the query on the client is redundant, since the results are serialized.

The JS client needs elaborate workarounds right now (including aliasing the search method) to be able to initialize the memory cache with already done queries.

Another use case is for prefilling the cache for e.g. empty or really popular queries.